style: 페이지 중복되는 컴포넌트 로직 useMovieList로분리후 재사용 완료
